+++ src/query/sql/src/planner/optimizer/rule/rewrite/rule_push_down_filter_join.rs
@@ -165,7 +165,7 @@ pub fn try_push_down_filter_join(
                         join.right_conditions.push(right.clone());
                         need_push = true;
                     }
-                } else if matches!(join.join_type, JoinType::Inner) {
+                } else if matches!(join.join_type, JoinType::Inner | JoinType::Cross) {
                     join.non_equi_conditions.push(predicate.clone());
                     need_push = true;
                 } else {
